# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentModuleKey = typeof StudentModuleKey[keyof typeof StudentModuleKey];


export const StudentModuleKey = {
  homework: 'homework',
  classwork: 'classwork',
  noticeboard: 'noticeboard',
  complaints: 'complaints',
} as const;

export interface StudentModuleActivityCursor {
  /**
     * Raw timestamp-without-time-zone text from the source event; used only as an ordering cursor.
     * @maxLength 26
     * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}(\.[0-9]{1,6})?$
     */
  createdAt: string;
  /** @minimum 1 */
  recordId: number;
}

export interface StudentModuleDotState {
  hasNewActivity: boolean;
  latestActivityCursor: StudentModuleActivityCursor | null;
}

export interface StudentModuleDotStateResponse {
  homework: StudentModuleDotState;
  classwork: StudentModuleDotState;
  noticeboard: StudentModuleDotState;
  complaints: StudentModuleDotState;
}

export interface StudentModuleSeenInput {
  module: StudentModuleKey;
  cursor: StudentModuleActivityCursor;
}

export interface StudentModuleSeenResponse {
  module: StudentModuleKey;
  markedSeen: boolean;
}

export type TeacherModuleKey = typeof TeacherModuleKey[keyof typeof TeacherModuleKey];


export const TeacherModuleKey = {
  noticeboard: 'noticeboard',
  complaints: 'complaints',
  leave: 'leave',
  approval_center: 'approval_center',
} as const;

export type TeacherModuleActivityCursorSource = typeof TeacherModuleActivityCursorSource[keyof typeof TeacherModuleActivityCursorSource];


export const TeacherModuleActivityCursorSource = {
  notice: 'notice',
  peer_report: 'peer_report',
  student_leave: 'student_leave',
  student_profile_submission: 'student_profile_submission',
  student_profile_photo: 'student_profile_photo',
} as const;

export interface TeacherModuleActivityCursor {
  /**
     * Raw source timestamp text used only for deterministic activity ordering.
     * @maxLength 26
     * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}(\.[0-9]{1,6})?$
     */
  createdAt: string;
  source: TeacherModuleActivityCursorSource;
  /** @minimum 1 */
  recordId: number;
}

export interface TeacherModuleDotState {
  hasNewActivity: boolean;
  latestActivityCursor: TeacherModuleActivityCursor | null;
}

export interface TeacherModuleDotStateResponse {
  noticeboard: TeacherModuleDotState;
  complaints: TeacherModuleDotState;
  leave: TeacherModuleDotState;
  approval_center: TeacherModuleDotState;
}

export interface TeacherModuleSeenInput {
  module: TeacherModuleKey;
  cursor: TeacherModuleActivityCursor;
}

export interface TeacherModuleSeenResponse {
  module: TeacherModuleKey;
  markedSeen: boolean;
}

export interface AdminTimetableContext {
  classes: string[];
  sections: string[];
  subjects: string[];
}

export type AdminCalendarContextClassSections = {[key: string]: string[]};

export interface AdminCalendarContext {
  classes: string[];
  sections: string[];
  classSections: AdminCalendarContextClassSections;
}

export interface AdminAttendanceContext {
  classes: string[];
  sections: string[];
  subjects: string[];
  attendanceTarget: number;
}

export interface AdminExamControllerContext {
  classes: string[];
  sections: string[];
  exam_types: string[];
}

export interface AdminNoticeboardContext {
  classes: string[];
  sections: string[];
}

export interface MobileTeacherFirstLoginPasswordChange {
  challengeToken: string;
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
  /** @minLength 6 */
  confirmPassword: string;
}

export const MobileTeacherFirstLoginPasswordChangedValue = {
  state: 'password_changed',
} as const;
export type MobileTeacherFirstLoginPasswordChanged = typeof MobileTeacherFirstLoginPasswordChangedValue;