# lib/api-zod/src/generated/types/teacherModuleActivityCursor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherModuleActivityCursorSource } from './teacherModuleActivityCursorSource';

export interface TeacherModuleActivityCursor {
  /**
     * Raw source timestamp text used only for deterministic activity ordering.
     * @maxLength 26
     * @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}(\.[0-9]{1,6})?$
     */
  createdAt: string;
  source: TeacherModuleActivityCursorSource;
  /** @minimum 1 */
  recordId: number;
}