# lib/api-zod/src/generated/types/teacherModuleDotState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherModuleActivityCursor } from './teacherModuleActivityCursor';

export interface TeacherModuleDotState {
  hasNewActivity: boolean;
  latestActivityCursor: TeacherModuleActivityCursor | null;
}