# lib/api-zod/src/generated/types/teacherModuleDotStateResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherModuleDotState } from './teacherModuleDotState';

export interface TeacherModuleDotStateResponse {
  noticeboard: TeacherModuleDotState;
  complaints: TeacherModuleDotState;
  leave: TeacherModuleDotState;
  approval_center: TeacherModuleDotState;
}