# lib/api-zod/src/generated/types/teacherModuleKey.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TeacherModuleKey = typeof TeacherModuleKey[keyof typeof TeacherModuleKey];


export const TeacherModuleKey = {
  noticeboard: 'noticeboard',
  complaints: 'complaints',
  leave: 'leave',
  approval_center: 'approval_center',
} as const;