# lib/api-zod/src/generated/types/teacherModuleSeenInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherModuleActivityCursor } from './teacherModuleActivityCursor';
import type { TeacherModuleKey } from './teacherModuleKey';

export interface TeacherModuleSeenInput {
  module: TeacherModuleKey;
  cursor: TeacherModuleActivityCursor;
}