# lib/api-zod/src/generated/types/teacherModuleSeenResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherModuleKey } from './teacherModuleKey';

export interface TeacherModuleSeenResponse {
  module: TeacherModuleKey;
  markedSeen: boolean;
}